--- packages/server/src/unstable-core-do-not-import/http/content-type/contentTypeHandlers.ts
@@ -1,10 +1,0 @@
-import { formDataContentTypeHandler } from './handlers/formData';
-import { jsonContentTypeHandler } from './handlers/json';
-
-export const contentTypeHandlers = {
-  list: [formDataContentTypeHandler, jsonContentTypeHandler],
-  /**
-   * Fallback handler if there is no match
-   */
-  fallback: jsonContentTypeHandler,
-};
