--- packages/server/src/unstable-core-do-not-import/http/content-type/handlers/json.ts
@@ -1,19 +1,0 @@
-import type { ContentTypeHandler } from '../../../../@trpc/server/http';
-
-export const jsonContentTypeHandler: ContentTypeHandler = {
-  async getInputs(req, searchParams) {
-    if (req.method === 'GET') {
-      const input = searchParams.get('input');
-      if (input === null) {
-        return undefined;
-      }
-      return JSON.parse(input);
-    }
-    return await req.json();
-  },
-  isMatch(req) {
-    return req.headers.get('content-type') == 'application/json';
-  },
-  batching: true,
-  transform: true,
-};
